Move RidePrinter from wikirider.py to view.py
